Cleanup object interfaces based on review suggestions
